refactor: reduce runtime dependency weight
fix: proto features

fix: layer

add missing features
