+++ Cargo.toml
@@ -99,7 +99,6 @@ tonic = "0.10.2"
 tower = "0.4.13"
 tower-http = { version = "0.4.0", features = ["trace"] }
 tracing = { version = "0.1.37", default-features = false }
-tracing-core = { version = "0.1.32", default-features = false }
 tracing-opentelemetry = "0.22.0"
 tracing-subscriber = { version = "0.3.16", default-features = false, features = [
   "registry",
