+++ runtime/src/alpha/mod.rs
@@ -11,7 +11,7 @@ use std::{
 use anyhow::Context;
 use async_trait::async_trait;
 use core::future::Future;
-use shuttle_common::{backends::trace::ExtractPropagationLayer, secrets::Secret};
+use shuttle_common::{extract_propagation::ExtractPropagationLayer, secrets::Secret};
 use shuttle_proto::runtime::{
     runtime_server::{Runtime, RuntimeServer},
     LoadRequest, LoadResponse, StartRequest, StartResponse, StopReason, StopRequest, StopResponse,
